fix error on DB shrink attempt while in 'heap' mode
chainbase does not allow the DB to be shrunk. If attempting this, the current expected behavior is to ignore the shrinkage request. However in heap mode w/o hugepages, there was an error where the wrong size gets plumbed through in this situation. Worst case can cause a crash on shutdown due to the mismatch.

Fix this error. Also add a warning telling users that the DB isn't being shrunk. Also add a small unittest.